requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
google-generativeai==0.3.2
apscheduler==3.10.4
pydantic==2.5.0
python-dateutil==2.8.2
pytz